feat (Edit)(controller): Atualização de nome de pasta e edição de arquivos de controle com criação de arquivos requisitados - @LuFi-1227
